fixed payout and settlement

+++ internal/dto/balance.go
@@ -1,9 +1,9 @@
 package dto
 
 type MerchantBalanceResponse struct {
-	MerchantID       int    `json:"merchant_id"`
-	Currency         string `json:"currency"`
-	PendingBalance   int64  `json:"pending_balance"`   // In kobo
-	AvailableBalance int64  `json:"available_balance"` // In kobo
-	TotalVolume      int64  `json:"total_volume"`      // In kobo
+	MerchantID       int     `json:"merchant_id"`
+	Currency         string  `json:"currency"`
+	PendingBalance   float64 `json:"pending_balance"`   // In currency units (e.g., NGN)
+	AvailableBalance float64 `json:"available_balance"` // In currency units (e.g., NGN)
+	TotalVolume      float64 `json:"total_volume"`      // In currency units (e.g., NGN)
 }
